requirements.txt
langgraph~=0.5.2
langchain~=0.3.26
pydantic~=2.11.7
python-dotenv~=1.1.1
langchain-core~=0.3.68
rich==14.0.0
pydantic-settings~=2.4.0
langchain_tavily==0.2.7
langchain-google-genai==2.1.7
langchain_google_vertexai==2.0.27
numexpr==2.11.0
flake8==7.3.0
wikipedia==1.4.0
waybackpy==3.0.6
requests~=2.32.0
beautifulsoup4==4.12.3
